% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustMeasurement.R
\name{adjustMeasurement}
\alias{adjustMeasurement}
\title{Adjusts measurement columns}
\usage{
adjustMeasurement(clams.list, msr.name)
}
\arguments{
\item{clams.list}{a CLAMS data or collection list}

\item{msr.name}{the measurement column name}
}
\value{
A CLAMS data or collection list. A CLAMS collection list contains
multiple CLAMS data lists. Each CLAMS data list contains two
elements:
\item{meta.data}{a list of meta data values}
\item{measurements}{a data frame of measurements}
}
\description{
Adjusts a measurement column from the measurements data frame of a
CLAMS data list or each CLAMS data list in a CLAMS collection list
using weight for normalization.
}
\details{
Each measurement data frame must contain an "ORIG.WEIGHT" and an
"ADJ.WEIGHT" column. The adjusted measurement is appended to the
data frame with name prefix "ADJ.".
}
\examples{
## Assign and load a CLAMS data directory
data.dir <- system.file("extdata", "Test", package="CLAMS")
clams.dir <- file.path(data.dir, "Collection-2015-01-06")
clams.coll <- loadClamsDir(clams.dir)

## Select columns and remove outliers
clams.coll <- selectColumns(clams.coll, do.remove.outliers=TRUE)

## Append original weight
clams.coll <- appendColumn(clams.coll, "orig.weight", c(23.8, 25.8, 22.6, 25.5, 25.5))

## Append adjusted weight in two intervals
clams.coll <- appendColumn(clams.coll, "adj.weight", c(23.5, 24.3, 22.6, 22.4, 24.9),
                           start.str="01/08/2015 06:00:00 AM", stop.str="01/09/2015 06:00:00 AM")
clams.coll <- appendColumn(clams.coll, "adj.weight", c(22.7, 24.2, 22.6, 25.1, 24.3),
                           start.str="01/11/2015 06:00:00 AM", stop.str="01/12/2015 06:00:00 AM")

## Adjust VO2 measurements
clams.coll <- adjustMeasurement(clams.coll, "VO2")
}

