% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignMeasurement.R
\name{alignMeasurement}
\alias{alignMeasurement}
\title{Aligns measurement columns}
\usage{
alignMeasurement(clams.list, msr.name, lbl.names = NULL, sel.name = NULL,
  sel.condition = NULL)
}
\arguments{
\item{clams.list}{a CLAMS data or collection list}

\item{msr.name}{the measurement column name}

\item{lbl.names}{the label column names}

\item{sel.name}{the selection column name}

\item{sel.condition}{the selection column single equality or
minimum, maximum pair condition}
}
\value{
A CLAMS measurement data frame containing the selected values of
the measurement and label columns from each CLAMS data list.
}
\description{
Aligns a measurement column from the measurements data frame of
each CLAMS data list in a CLAMS collection list to a common time
sequence in seconds.
}
\details{
Optionally includes label columns which may be required for
plotting. By default, label columns are not included.

Optionally selects a subset of the measurements based on a
selection named column single equality or minimum, maximum pair
column value condition. By default, all measurements are inlcuded.
}
\examples{
\dontrun{
clams.msr <- alignMeasurement(clams.coll, "VO2")
clams.msr <- alignMeasurement(clams.coll, "VO2", sel.name="LIGHT", sel.condition=TRUE)
clams.msr <- alignMeasurement(clams.coll, "VO2", sel.name="XTOT", sel.condition=c(250, 750))
clams.msr <- alignMeasurement(clams.coll, "VO2", "TEMP")
clams.msr <- alignMeasurement(clams.coll, "VO2", c("LIGHT", "DARK"))
}
}

