% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analyzeMeasurement.R
\name{analyzeMeasurement}
\alias{analyzeMeasurement}
\title{Analyzes measurement columns}
\usage{
analyzeMeasurement(clams.coll, msr.name, mdl.formula = NULL,
  sel.name = NULL, sel.condition = NULL, agg.functions = NULL)
}
\arguments{
\item{clams.coll}{a CLAMS collection list}

\item{msr.name}{the measurement column name}

\item{mdl.formula}{the character formula specifying the model}

\item{sel.name}{the selection column name}

\item{sel.condition}{the selection column single equality or
minimum, maximum pair condition}

\item{agg.functions}{the functions used to aggregate the
measurements}
}
\value{
A CLAMS analysis list containing three elements:
\item{name}{the name of the analyzed measurement}
\item{data}{a data frame containing the aligned and aggregated
measurements}
\item{model}{an analysis of variance model}
\item{effect}{an effects object}
}
\description{
Aligns a measurement column from the measurements data frame of
each CLAMS data list in a CLAMS collection list, then fits an
analysis of variance model, and computes a corresponding effects
object.
}
\details{
Optionally specifies the model using a formula. By default, the
formula is specified as:
  paste(msr.name, "~ GENOTYPE + ADJ.WEIGHT")

Optionally selects a subset of the measurements based on a
selection named column single equality or minimum, maximum pair
column value condition. By default, all measurements are included.

Optionally specifies the function used to aggregate the
measurements. By default, the \code{mean} is used to aggregate
measurements of oxygen consumption ("VO2"), carbon monoxide
production ("VCO2"), respiratory exchange ratio ("RER"), and body
temperature ("BODY.TEMP"), and the \code{sum} is used to
aggregate heat production ("HEAT"), accumulated food consumption
("FEED1.ACC"), and total ("XTOT") and ambient ("XAMB")
activity. The default can be overridden for any measurement by
providing a list with an element equal to the desired function and
named for the measurement.

Note that the \code{mean} is used to aggregate all other numeric
column values, and \code{unique} is used to aggregate character
column values.
}
\examples{
## Assign and load a CLAMS data directory
data.dir <- system.file("extdata", "Test", package="CLAMS")
clams.dir <- file.path(data.dir, "Collection-2015-01-06")
clams.coll <- loadClamsDir(clams.dir)

## Select columns and remove outliers
clams.coll <- selectColumns(clams.coll, do.remove.outliers=TRUE)

## Append a column indicating genotype, temperature, original weight,
## and adjusted weight
clams.coll <- appendColumn(clams.coll, "genotype", c("KO", "WT", "KO", "KO", "WT"))
clams.coll <- appendColumn(clams.coll, "temp.22", TRUE,
                           start.str="01/08/2015 06:00:00 AM", stop.str="01/09/2015 06:00:00 AM")
clams.coll <- appendColumn(clams.coll, "orig.weight", c(23.8, 25.8, 22.6, 25.5, 25.5))
clams.coll <- appendColumn(clams.coll, "adj.weight", c(23.5, 24.3, 22.6, 22.4, 24.9),
                           start.str="01/08/2015 06:00:00 AM", stop.str="01/09/2015 06:00:00 AM")

## Analyze heat production
clams.analysis <- analyzeMeasurement(clams.coll, "HEAT", sel.name="TEMP.22", sel.condition=TRUE)
}

