% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/appendColumn.R
\name{appendColumn}
\alias{appendColumn}
\title{Appends a named column value in a time interval}
\usage{
appendColumn(clams.list, col.name, col.value, start.str = "", stop.str = "",
  is.daily = FALSE)
}
\arguments{
\item{clams.list}{a CLAMS data or collection list}

\item{col.name}{the column name}

\item{col.value}{the column value}

\item{start.str}{the start time string in format "\%I:\%M:\%S \%p"
or "\%m/\%d/\%Y \%I:\%M:\%S \%p"}

\item{stop.str}{the stop time string in format "\%I:\%M:\%S \%p" or
"\%m/\%d/\%Y \%I:\%M:\%S \%p"}

\item{is.daily}{flag to indicate whether the time interval recurs
daily, or not}
}
\value{
A CLAMS data or collection list. A CLAMS collection list contains
multiple CLAMS data lists. Each CLAMS data list contains two
elements:
\item{meta.data}{a list of meta data values}
\item{measurements}{a data frame of measurements}
}
\description{
Appends a named column value in a time interval to the measurements
data frame of a CLAMS data list or of each CLAMS data list in a
CLAMS collection list.
}
\details{
Translates the column name to upper case characters.

The column value must have length one or the number of CLAMS data
lists. If length one, the same column value is appended to each
data list. Otherwise, the column value corresponding to each data
list is appended.

Assignment in the time interval requires that the CLAMS measurement
data frame contain a date-time column. One is prepended, if
needed. The start and stop time strings are optional. If absent,
the start time will correspond to the start time of the
measurements, and stop time will correspond to the stop time of the
measurements.

By default, the time interval does not recur daily. If
\code{is.daily} is \code{TRUE}, the start and stop time strings
must contain only time of day, and not date.
}
\examples{
## Assign and load a CLAMS data directory, and select default columns
data.dir <- system.file("extdata", "Test", package="CLAMS")
clams.dir <- file.path(data.dir, "Collection-2013-01-15")
clams.coll <- loadClamsDir(clams.dir)
clams.coll <- selectColumns(clams.coll)

## Append a column indicating daily periods of illumination
clams.coll <- appendColumn(clams.coll, "light", TRUE,
                           start.str="06:00:00 AM", stop.str="06:00:00 PM", is.daily=TRUE)
}

