% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadClamsDir.R
\name{loadClamsDir}
\alias{loadClamsDir}
\title{Load a collection of data files}
\usage{
loadClamsDir(data.dir)
}
\arguments{
\item{data.dir}{the directory containing CLAMS data files}
}
\value{
A CLAMS collection list containing a CLAMS data list for each CLAMS
data file. A CLAMS data list contains two elements:
\item{meta.data}{a list of meta data values}
\item{measurements}{a data frame of measurements}
}
\description{
Loads a collection of CLAMS data files.
}
\details{
Loads each CLAMS data file contained in a directory.
}
\examples{
\dontrun{
clams.coll <- loadClamsDir("./Data/2013-01-15.0101")
}
}

