% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadClamsFile.R
\name{loadClamsFile}
\alias{loadClamsFile}
\title{Loads a data file}
\usage{
loadClamsFile(input.file)
}
\arguments{
\item{input.file}{the file containing the CLAMS data}
}
\value{
A CLAMS data list containing two elements:
\item{meta.data}{a list of meta data values}
\item{measurements}{a data frame of measurements}
}
\description{
Loads a CLAMS data file.
}
\details{
The CLAMS data file is processed line by line in one of three
modes: "META", "DATA", and "EVENTS". These modes correspond to the
sections of the file. Data lines are written to a file in the same
directory and with the same name as the input file but with the
word "Extracted" added prior to the file extension.

While in mode "DATA" lines deemed to be invalid are written to a
file in the same directory and with the same name as the input file
but with the word "Rejected" added prior to the file extension. For
example, lines which contain three consecutive measurements which
are reported as zero to three significant digits are rejected.
}
\examples{
## Assign a CLAMS data file
data.dir <- system.file("extdata", "Test", package="CLAMS")
clams.file <- file.path(data.dir, "Collection-2013-01-15", "2013-01-15.0101.CSV")

## Load a CLAMS data file
clams.data <- loadClamsFile(clams.file)
}

