% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMeasurement.R
\name{plotMeasurement}
\alias{plotMeasurement}
\title{Plots aligned measurement columns}
\usage{
plotMeasurement(clams.msr, msr.label, output.file, agg.names, agg.labels,
  agg.function = mean, con.names = NULL, con.labels = NULL,
  do.filter = TRUE, do.error.bars = FALSE)
}
\arguments{
\item{clams.msr}{aligned measurement columns}

\item{msr.label}{measurement column labels}

\item{output.file}{file to contain plot output}

\item{agg.names}{aggregration column names}

\item{agg.labels}{aggregration column labels}

\item{agg.function}{aggregration function}

\item{con.names}{test condition column names}

\item{con.labels}{test condition plot labels}

\item{do.filter}{flag to filter measurements, or not}

\item{do.error.bars}{flag to plot error bars, or not}
}
\value{
A boolean indicating success, or not.
}
\description{
Plots aligned measurement columns from the measurements data frame
of each CLAMS data list in a CLAMS collection list as a function of
a common time sequence in seconds.
}
\details{
Places PDF commands needed for the plot in output.file.

Aggregates named measurement columns as labeled columns using an
aggregation function, by default using \code{mean}. Optionally
plots and labels named test conditions. Note that measurement
column and test condition labels may be character or expression
vectors.

Optionally filters the measurements by applying a trailing one hour
average. By default, measurements are filtered. Optionally plots
error bars. By default, error bars are not plotted.
}
\examples{
##  Assign and load a CLAMS data directory
data.dir <- system.file("extdata", "Test", package="CLAMS")
clams.dir <- file.path(data.dir, "Collection-2014-12-19")
clams.coll <- loadClamsDir(clams.dir)

## Select columns and remove outliers
clams.coll <- selectColumns(clams.coll, do.remove.outliers=TRUE)

## Append test conditions
clams.coll <- appendColumn(clams.coll, "LIGHT", TRUE,
                           start.str="06:00:00 AM", stop.str="06:00:00 PM", is.daily=TRUE)
clams.coll <- appendColumn(clams.coll, "DARK", TRUE,
                           start.str="06:00:00 PM", stop.str="06:00:00 AM", is.daily=TRUE)
clams.coll <- appendColumn(clams.coll, "TEMP.30", TRUE,
                           start.str="12/21/2014 6:00:00 AM", stop.str="12/22/2014 6:00:00 AM")
clams.coll <- appendColumn(clams.coll, "TEMP.22", TRUE,
                           start.str="12/24/2014 6:00:00 AM", stop.str="12/25/2014 6:00:00 AM")

## Align oxygen consumption measurments for each specimen and append
## test conditions
clams.msr <- alignMeasurement(clams.coll, "VO2", c("LIGHT", "DARK", "TEMP.30", "TEMP.22"))

## Plot aligned measurmeents, aggregating and labeling named columns,
## labeling test conditions, filtering measurements, and plotting
## error bars
plotMeasurement(clams.msr, "Oxygen Consumption, ml/kg/hr", "VO2.pdf",
                agg.names=list(c("CRE.NEG.NH", "CRE.NEG.HL", "CRE.NEG.RL"),
                               c("CRE.POS.HR.2", "CRE.POS.HR.4", "CRE.POS.2R")),
                agg.labels=c("CRE.NEG", "CRE.POS"),
                con.names=c("TEMP.30", "TEMP.22"),
                con.labels=expression(paste("30", degree, "C"), paste("22", degree, "C")),
                do.filter=TRUE, do.error.bars=TRUE)
}

