% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/selectColumns.R
\name{selectColumns}
\alias{selectColumns}
\title{Selects named columns}
\usage{
selectColumns(clams.list, col.names = c(), do.remove.outliers = FALSE)
}
\arguments{
\item{clams.list}{a CLAMS data or collection list}

\item{col.names}{the column names to select}

\item{do.remove.outliers}{flag to remove outliers, or not}
}
\value{
A CLAMS data or collection list. A CLAMS collection list contains
multiple CLAMS data lists. Each CLAMS data list contains two
elements:
\item{meta.data}{a list of meta data values}
\item{measurements}{a data frame of measurements}
}
\description{
Selects named columns from the measurements data frame of a
CLAMS data list or of each CLAMS data list in a CLAMS collection
list. Optionally, removes outliers.
}
\details{
The vector of measurement column names is optional. If absent, the
default selection of columns includes: "INTERVAL", "CHAN",
"DATE.TIME", "VO2", "VCO2", "HEAT", "RER", "FEED1.ACC", "XTOT",
"XAMB", and "BODY.TEMP".

By default outliers are not removed. If \code{do.remove.outliers}
is \code{TRUE}, outliers, including non-physical values, and values
greater than three standard deviations from the mean, are
removed. Non-physical values include values of VO2, VCO2, and HEAT
less than or equal to zero, RER less than 0.5 or more than 1.2, and
XTOT and XAMB less than zero.
}
\examples{
## Assign and load a CLAMS data file
data.dir <- system.file("extdata", "Test", package="CLAMS")
clams.file <- file.path(data.dir, "Collection-2013-01-15", "2013-01-15.0101.CSV")
clams.data <- loadClamsFile(clams.file)

## Select default columns
clams.data <- selectColumns(clams.data)

## Assign and load a CLAMS data directory
clams.dir <- file.path(data.dir, "Collection-2013-01-15")
clams.coll <- loadClamsDir(clams.dir)

## Select default columns, a single column, and a single column with
## outliers removed
clams.coll <- selectColumns(clams.coll)
clams.coll <- selectColumns(clams.coll, col.names=c("VO2"))
clams.coll <- selectColumns(clams.coll, col.names=c("VO2"), do.remove.outliers=TRUE)
}

