% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMeasurement.R
\name{plotMeasurement}
\alias{plotMeasurement}
\title{Plots aligned measurement columns}
\usage{
plotMeasurement(clams.msr, msr.label, output.file, agg.names, agg.labels,
  agg.function = mean, con.names = NULL, con.labels = NULL,
  do.filter = TRUE, do.error.bars = FALSE)
}
\arguments{
\item{clams.msr}{aligned measurement columns}

\item{msr.label}{measurement column labels}

\item{output.file}{file to contain plot output}

\item{agg.names}{aggregration column names}

\item{agg.labels}{aggregration column labels}

\item{agg.function}{aggregration function}

\item{con.names}{test condition column names}

\item{con.labels}{test condition plot labels}

\item{do.filter}{flag to filter measurements, or not}

\item{do.error.bars}{flag to plot error bars, or not}
}
\value{
A boolean indicating success, or not.
}
\description{
Plots aligned measurement columns from the measurements data frame
of each CLAMS data list in a CLAMS collection list as a function of
a common time sequence in seconds.
}
\details{
Places PDF commands needed for the plot in output.file.

Aggregates named measurement columns as labeled columns using an
aggregation function, by default using \code{mean}. Optionally
plots and labels named test conditions. Note that measurement
column and test condition labels may be character or expression
vectors.

Optionally filters the measurements by applying a trailing one hour
average. By default, measurements are filtered. Optionally plots
error bars. By default, error bars are not plotted.
}
\examples{
\dontrun{
plotMeasurement(clams.msr, "Oxygen Consumption, ml/kg/hr", "VO2.pdf",
                agg.names=list(c("WT1", "WT2"), c("KO1", "KO2")),
                agg.labels=c("WT", "KO"),
                con.names=c("TEMP.23", "TEMP.04"),
                con.labels=expression(paste("23", degree, "C"), paste("4", degree, "C")),
                do.filter=TRUE, do.error.bars=TRUE)
plotMeasurement(clams.msr, "Oxygen Consumption, ml/kg/hr", "VO2.pdf",
                agg.names=list(c("WT1", "WT2"), c("KO1", "KO2")),
                agg.labels=c("WT", "KO"), agg.function=sum,
                con.names=c("TEMP.23", "TEMP.04"),
                con.labels=expression(paste("23", degree, "C"), paste("4", degree, "C")),
                do.filter=TRUE, do.error.bars=TRUE)
}
}

